requirements.txt
asgiref==3.9.0
cfgv==3.4.0
distlib==0.3.9
dj-database-url==3.0.1
Django==5.2.4
django-extensions==4.1
djangorestframework==3.15.2
filelock==3.18.0
identify==2.6.12
nodeenv==1.9.1
platformdirs==4.3.8
pre_commit==4.2.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
PyYAML==6.0.2
sqlparse==0.5.3
virtualenv==20.31.2
